fix shift error when showing error
